+++ src/TimeHacker.Application/Models/Input/Tasks/InputFixedTaskModel.cs
@@ -5,7 +5,7 @@
         public string Name { get; set; }
         public string? Description { get; set; }
         public IEnumerable<int> CategoryIds { get; set; }
-        public uint Priority { get; set; }
+        public byte Priority { get; set; }
         public string StartTimestamp { get; set; }
         public string EndTimestamp { get; set; }
     }
